refactor: split config.py into a package by what each model configures
667 lines, 61 fields and 12 validators in one module, about to take another six
fields and four validators when the board goes continuous — a scale, a base
radius, an engagement range, a sample step, polygon footprints and polygon
objectives, each landing in a different class.

Now `types/config/`: `battle` (turn order, opponent, mission), `entities`
(weapons, models, objectives), `terrain` (pieces, named maps, the generator),
`env` (`WargameEnvConfig`), and `_validation` for the checks they share. The
dependency runs one way — `env` imports the rest and nothing imports `env`.

The package boundary is deliberately invisible: `__init__.py` re-exports every
public name, so all 72 importing files are untouched. That is the property that
makes this safe, and it is verified rather than assumed — `git diff --name-only`
against the previous commit shows changes under `types/config/` and nowhere
else.

876 tests pass, mypy clean, and the 8-episode digest is unchanged.
